reorg 2
Single inlcude to use mode, Loop now called AncientArcher, controls and display parented to AA class. library usage with a single include of AncientArcher.h
